+++ composeApp/src/commonMain/kotlin/presentation/authentication/AuthenticationContract.kt
@@ -17,7 +17,7 @@ class AuthenticationContract {
         fun goToCompleteProfileWithPhone(phone: String)
         fun showUserLocation(place: Place)
         fun goToConnectVendor(user: User)
-        fun onCompleteProfileDone(country: String, profileId: Int)
+        fun onCompleteProfileDone(country: String, profileId: Long)
         fun onProfileValidationStarted()
         fun onProfileValidationEnded()
         fun onCompleteProfileStarted()
